requirements.txt
speechrecognition
pyaudio
python-docx
pyperclip
